Fixed other backtrace bug related to extend_flag and proper gap-to position
Was setting previ and prevj based on self.backPos[i,j] in all cases. However, when extend_flag is true we need to use the connected position from the associated T node rather than the matrix position's chosen backPos. This is because backPos doesn't know that the location the back-trace came from requires a gap continuation.